Refactor:  split config and nodes
